Fix: UI related issues RMNCH - PNC Flow.
    - Pregnancy lmp date update from medical review.
    - Different login dialogue.
    - Query changes on unsynced household data count.
    - MDT-14402.

+++ Spice-SL/app/src/main/java/com/medtroniclabs/spice/data/offlinesync/model/PregnancyDetails.kt
@@ -2,7 +2,7 @@ package com.medtroniclabs.spice.data.offlinesync.model
 
 data class PregnancyDetails(
     val householdMemberId: String,
-    val ancVisitNo: Int?,
+    val ancVisitNo: Long?,
     val lastMenstrualPeriod: String?,
     val pncVisitNo: Int?,
     val dateOfDelivery: String?,
